+++ src/lib/data-collector.js
@@ -1,7 +1,7 @@
 'use strict';
 
 var constants = require('../constants');
-var assets = require('./assets');
+var assets = require('@braintree/asset-loader');
 var Promise = require('./promise');
 
 function DataCollector(config) {